% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b}
\alias{b}
\title{@title  Example parameter b}
\format{
A \code{vector}
\describe{
   \item{b}{a vector with values for b, generated randomly,
   range from 200 to 300}
}
}
\source{
Random generation
}
\usage{
b
}
\description{
@description Data generated randomly
}
\author{
CGE
}
\keyword{datasets}
