% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bstat.R
\name{Bstat}
\alias{Bstat}
\title{Basic statistical measures of a mathematical model results}
\usage{
Bstat(out_set)
}
\arguments{
\item{out_set}{matrix of dimensions n x t, where n equals the number of
runs and t is equal to the number of temporary steps.}
}
\value{
a data frame of dimensions t x 6, here t is the number of temporary steps
and each column corresponds to a statistical measure: mean, variance,
skewness, kurtosis and excess kurtosis.
}
\description{
This function calculates the mean, variance, skewness, kurtosis and
excess kurtosis of a model output, this output can be given for
different temporal periods (days, months or years).
}
\examples{
data("out_set")
data_Bstat <- Bstat(out_set)

}
\author{
Camila Garcia-Echeverri <cagarciae@unal.edu.co> \cr

Hydrodynamics of the natural media research group - HYDS
National University of Colombia -  Bogota
}
