% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c}
\alias{c}
\title{@title  Example parameter c}
\format{
A \code{vector}
\describe{
   \item{c}{a vector with values for c, generated randomly,
   range from 0.1 to 0.2}
}
}
\source{
Random generation
}
\usage{
c
}
\description{
@description Data generated randomly
}
\author{
CGE
}
\keyword{datasets}
