% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_Bstat}
\alias{data_Bstat}
\title{@title  First four conditional moments of example data}
\format{
A \code{data.frame}
\describe{
   \item{data_Bstat}{a data frame of dimensions t x 6}
}
}
\source{
Function Bstat
}
\usage{
data_Bstat
}
\description{
@description Data generated with the example of the function Cond_Moments
}
\author{
Camila Garcia-Echeverri
}
\keyword{datasets}
