% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_set}
\alias{parameters_set}
\title{@title  Set of parameters randomly generated}
\format{
A \code{matrix}
\describe{
   \item{parameters_set}{a matrix of dimensions 500 x 10 (n x pp),runs of the model x
   number of parameters}
}
}
\usage{
parameters_set
}
\description{
@description  It contains 10 parameters
}
\references{
Arenas-Bautista, M. C. (2020). Integration of Hydrological and Economical
Aspects for Water Management in Tropical Regions. Case Study: Middle Magdalena Valley, Colombia.
National University of Colombia.
}
\keyword{datasets}
