% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp_names}
\alias{pp_names}
\title{@title  Example - parameters names}
\format{
A \code{value}
\describe{
   \item{pp_names}{a vector of characters}
}
}
\usage{
pp_names
}
\description{
@description 10 parameters names.
}
\references{
Arenas-Bautista, M. C. (2020). Integration of Hydrological and Economical
Aspects for Water Management in Tropical Regions. Case Study: Middle Magdalena Valley, Colombia.
National University of Colombia.
}
\author{
CGE
}
\keyword{datasets}
