% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess}
\alias{preProcess}
\title{Preprocessing entry data}
\usage{
preProcess(x, y)
}
\arguments{
\item{#ENTRY}{2  HASTA EL NUMERO DE ENTRADAS}
}
\value{
QUÉ DEVUELVE?
}
\description{
This function XXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
}
\examples{



}
\author{
Camila Garcia-Echeverri <cagarciae@unal.edu.co> \cr
Maria Cristina Areas-Bautista <mcarenasb@unal.edu.co> \cr
Leonardo David Donado <lddonadog@unal.edu.co> \cr


Hydrodynamics of the natural media research group - HYDS
Universidad Nacional de Colombia - sede Bogota
}
