% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{Save GSA results}
\usage{
save_results(SOBOL, SOBOL_total, amae, amav, amar, amak)
}
\arguments{
\item{SOBOL}{SOBOL index}

\item{SOBOL_total}{SOBOL_total}

\item{amae}{AMAE index}

\item{amav}{AMAV index}

\item{amar}{AMAR index}

\item{amak}{AMAK index}
}
\description{
This function helps to save the results in .csv format
}
\author{
Camila Garcia-Echeverri <cagarciae@unal.edu.co> \cr

Hydrodynamics of the natural media research group - HYDS
Universidad Nacional de Colombia - sede Bogota
}
