% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z}
\alias{z}
\title{@title  Example function results}
\format{
A \code{vector}
\describe{
   \item{z}{a vector with alpha1 values for each GRU}
}
}
\source{
Synthetic generation
}
\usage{
z
}
\description{
@description Data generated with an example funtion
}
\examples{
\dontrun{
data(ExampleFUN)
summary(ExampleFUN)
}

}
\author{
CGE
}
\keyword{datasets}
