% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a}
\alias{a}
\title{@title  Example parameter a}
\format{
A \code{vector}
\describe{
   \item{a}{a vector with values for a, generated randomly,
   range from 1 to 10}
}
}
\source{
Random generation
}
\usage{
a
}
\description{
@description Data generated randomly
}
\author{
CGE
}
\keyword{datasets}
