% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CM}
\alias{CM}
\title{@title  First four conditional moments of example data}
\format{
A \code{list}
\describe{
   \item{CM}{A list of arrays, each array has dimensions of steps, t, p}
}
}
\source{
Function Conditional_Moments
}
\usage{
CM
}
\description{
@description Data generated with the example of the function Cond_Moments
}
\author{
CGE
}
\keyword{datasets}
