% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAtool.R
\name{GSAtool}
\alias{GSAtool}
\title{Global Sensibility Analisys tool}
\usage{
GSAtool(parameters_set, out_set, pp_names, steps = 100)
}
\arguments{
\item{parameters_set}{matrix of dimensions n x pp, where n is the
number of runs and pp is the number of parameters.}

\item{out_set}{matrix of dimensions n x t, where n is the number of
runs and t is the number of tempory steps.}

\item{pp_names}{a strings vector with the names of the parameters of the model}

\item{steps}{number of divisions of the parametric range}
}
\value{
a list containing two matrices. The fisrt contains the first order sobol, the
second sobol_total
}
\description{
This function xxx
}
\examples{
data("parameters_set", "out_set", "pp_names")
GSA_results <- GSAtool(parameters_set, out_set, pp_names, steps = 20)

}
\references{
Dell’Oca, A., Riva, M., & Guadagnini, A. (2017). Moment-based metrics for global sensitivity
analysis of hydrological systems. Hydrology and Earth System Sciences, 21(12), 6219–6234.
 https://doi.org/10.5194/hess-21-6219-2017

Sobol, I. M. (2001). Global sensitivity indices for nonlinear mathematical models
and their Monte Carlo estimates. Mathematics and Computers in Simulation, 55(1–3),
271–280. https://doi.org/10.1016/S0378-4754(00)00270-6
}
\author{
Camila Garcia-Echeverri <cagarciae@unal.edu.co> \cr
Maria Cristina Areas-Bautista <mcarenasb@unal.edu.co> \cr


Hydrodynamics of the natural media research group - HYDS
Universidad Nacional de Colombia - sede Bogota
}
