% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{out_set}
\alias{out_set}
\title{@title  Results of a sample model}
\format{
A \code{matrix}
\describe{
   \item{out_set}{a matrix of dimensions n x t, runs of the model x
   temporary steps}
}
}
\source{
Random generation
}
\usage{
out_set
}
\description{
@description Output generated with an example funtion
 x <- z*exp(a)+b*log(c)
}
\keyword{datasets}
