% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out_set}
\alias{out_set}
\title{@title  Results of a sample model}
\format{
A \code{matrix}
\describe{
   \item{out_set}{a matrix of dimensions 500 x 365 (pp x t), runs of the model x
   temporary steps (365 days)}
}
}
\usage{
out_set
}
\description{
@description Output generated with an example mathematical model.
}
\references{
Arenas-Bautista, M. C. (2020). Integration of Hydrological and Economical
Aspects for Water Management in Tropical Regions. Case Study: Middle Magdalena Valley, Colombia.
National University of Colombia.
}
\keyword{datasets}
