% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_set}
\alias{parameters_set}
\title{@title  Set of parameters randomly generated}
\format{
A \code{matrix}
\describe{
   \item{parameters_set}{a matrix of dimensions n x 3, runs of the model x
   number of parameters}
}
}
\source{
Maria Crstina thesis XXX
}
\usage{
parameters_set
}
\description{
@description  It contains the parameters a,b,c.
}
\keyword{datasets}
