% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp_names}
\alias{pp_names}
\title{@title  Example - parameters names}
\format{
A \code{value}
\describe{
   \item{pp_names}{a vector of characters}
}
}
\source{
Maria Crstina thesis XXX
}
\usage{
pp_names
}
\description{
@description Parameters names
}
\author{
CGE
}
\keyword{datasets}
