% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z}
\alias{z}
\title{@title  Example input z}
\format{
A \code{vector}
\describe{
   \item{z}{a vector with values for z, generated randomly}
}
}
\source{
Random generation
}
\usage{
z
}
\description{
@description Data generated randomly
}
\keyword{datasets}
